Restructured scriptsystem and scriptcomponent
Moved some functionality to the component. It's no longer a POD, but it made more sense tbh. All script components have a JS engine each, this is much better for performance. ScriptSystem as a global object in the JS engines also makes more sense now. App and the script widget are still the only ones directly accessing the script components (which is what we want):