Changed getComponent to use binary search
As our lists are always ordered anyway, it's no real reason not to use binary search. Our implementation uses std::lower_bound which is a binary search that finds the first not greater element. In our implementation we therefore also need to check if the iterator returned actually equals to the one we're looking for.